latest code added

+++ src/main/java/com/hussain/dto/FundTransferRequestDTO.java
@@ -1,7 +1,5 @@
 package com.hussain.dto;
 
-import java.math.BigDecimal;
-
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
@@ -25,7 +23,7 @@ public class FundTransferRequestDTO  {
 
     @NotNull(message = "Amount is required")
     @Positive(message = "Amount must be greater than zero")
-    private BigDecimal amount;
+    private Double amount;
 
     private String comments;
